filter_images/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0
open_clip_torch>=2.20.0
matplotlib>=3.7.0

